Build/Master/requirements.txt
setuptools==41.0.0
# apache-airflow[kubernetes]==2.0.1  # unresolvable conflict with alembic_utils (alembic<1.2 vs alembic>=1.5.7)
alembic==1.7.3
alembic_utils==0.7.2
arnparse==0.0.2
boto3==1.18.12
bs4==0.0.1
coverage==5.1
croniter==0.3.37
cryptography==3.4.8
dask==2021.10.0
docker==4.2.1
doit==0.31.1
endesive==2.0.7
Flask==1.1.1
feather-format==0.4.0
gevent==1.4.0
gremlinpython==3.5.2
gunicorn==19.10.0
guppy3==3.1.1
imbalanced-learn==0.6.2
ipython==7.32.0
JayDeBeApi==1.1.1
jinja2==2.11.3
jpype1==0.6.3
json5==0.8.5
jupyter==1.0.0
lightgbm==2.3.1
lxml==4.6.5
MarkupSafe==1.1.1  # pinned for compatibility with jinja2
marshmallow==3.10.0
matplotlib==3.1.3
mistune==2.0.1  # pinned for compatibility
mock==4.0.2
nameparser==1.0.6
notebook==6.4.8
numpy==1.18.1
openpyxl==3.0.7
pandas==1.0.1
paradag==1.2.0
pika==1.2.0
pip==22.0.4
pipenv==2018.5.18
psycopg2-binary==2.8.5
pydeps==1.9.0
pygments==2.8.1
pylint==2.7.0
pymysql==1.0.2
pyodbc==4.0.28; platform_system=="Windows"
PyPDF2==1.26.0
pysftp==0.2.9
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pywin32==223; platform_system=="Windows"
PyYAML==5.4
requests==2.26.0
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
sqlalchemy==1.3.18
tensorflow==2.1.0; platform_system=="Windows"
twine==4.0.0
urllib3==1.26.5
wheel==0.37.1
xgboost==0.90
xlrd==2.0.1  # pinned for compatibility
XlsxWriter==1.3.7
